rename components to - too
