+++ src/javascript/app_2/App/Components/Elements/DataTable/index.js
@@ -1,1 +1,1 @@
-export default from './data_table.jsx';
+export default from './data-table.jsx';
