+++ src/javascript/app_2/App/Components/Elements/SettingsDialog/settings_control.jsx
@@ -1,6 +1,6 @@
 import PropTypes    from 'prop-types';
 import React        from 'react';
-import SwitchButton from '../switch_button.jsx';
+import SwitchButton from '../switch-button.jsx';
 
 const SettingsControl = ({
     children,
