+++ src/javascript/app_2/App/Components/Elements/ToggleButton/__tests__/toggle-button.spec.js
@@ -2,7 +2,7 @@ import { expect }   from 'chai';
 import { shallow }  from 'enzyme';
 import React        from 'react';
 import { fake }     from 'sinon';
-import ToggleButton from '../toggle_button.jsx';
+import ToggleButton from '../toggle-button.jsx';
 import Button       from '../../../Form/button.jsx';
 
 describe('<ToggleButton />', () => {
