+++ src/javascript/app_2/App/Components/Elements/ToggleButton/index.js
@@ -1,2 +1,2 @@
-export ToggleButton      from './toggle_button.jsx';
-export ToggleButtonGroup from './toggle_button_group.jsx';
+export ToggleButton      from './toggle-button.jsx';
+export ToggleButtonGroup from './toggle-button-group.jsx';
