+++ src/javascript/app_2/App/Components/Routes/route_with_sub_routes.jsx
@@ -5,7 +5,7 @@ import {
 import { redirectToLogin } from '_common/base/login';
 import routes              from 'Constants/routes';
 import GTM                 from 'Utils/gtm';
-import LoginPrompt         from '../Elements/login_prompt.jsx';
+import LoginPrompt         from '../Elements/login-prompt.jsx';
 import { default_title }   from '../../Constants/app_config';
 
 const RouteWithSubRoutes = route => {
