+++ src/javascript/app_2/App/Containers/AccountSwitcher/account-switcher.jsx
