+++ src/javascript/app_2/App/Containers/AccountSwitcher/index.js
@@ -1,1 +1,1 @@
-export * from './account_switcher.jsx';
+export * from './account-switcher.jsx';
