+++ src/javascript/app_2/App/app.jsx
@@ -3,10 +3,10 @@ import React                       from 'react';
 import { BrowserRouter as Router } from 'react-router-dom';
 import getBaseName                 from 'Utils/URL/base_name';
 import { MobxProvider }            from 'Stores/connect';
-import ErrorBoundary               from './Components/Elements/Errors/error_boundary.jsx';
+import ErrorBoundary               from './Components/Elements/Errors/error-boundary.jsx';
 import PositionsDrawer             from './Components/Elements/PositionsDrawer';
 import { POSITIONS }               from './Components/Elements/ToastMessage';
-import ToastMessage                from './Containers/toast_message.jsx';
+import ToastMessage                from './Containers/toast-message.jsx';
 import AppContents                 from './Containers/Layout/app_contents.jsx';
 import Footer                      from './Containers/Layout/footer.jsx';
 import Header                      from './Containers/Layout/header.jsx';
