+++ src/javascript/app_2/Modules/Portfolio/Containers/portfolio.jsx
@@ -6,7 +6,7 @@ import { withRouter }                 from 'react-router-dom';
 import DataTable                      from 'App/Components/Elements/DataTable';
 import { getContractPath }            from 'App/Components/Routes/helpers';
 import { connect }                    from 'Stores/connect';
-import CardList                       from '../Components/card_list.jsx';
+import CardList                       from '../Components/card-list.jsx';
 import EmptyPortfolioMessage          from '../Components/empty_portfolio_message.jsx';
 import { getTableColumnsTemplate }    from '../Constants/data_table_constants';
 import Loading                        from '../../../../../templates/_common/components/loading.jsx';
