+++ src/javascript/app_2/Modules/Trading/Components/Form/TradeParams/allow-equals.jsx
