+++ src/javascript/app_2/Modules/Trading/Components/Form/TradeParams/amount.jsx
@@ -10,7 +10,7 @@ import {
 import Dropdown                  from 'App/Components/Form/DropDown';
 import Fieldset                  from 'App/Components/Form/fieldset.jsx';
 import InputField                from 'App/Components/Form/input_field.jsx';
-import AllowEquals               from './allow_equals.jsx';
+import AllowEquals               from './allow-equals.jsx';
 
 const Amount = ({
     amount,
