+++ src/javascript/app_2/Modules/Trading/Components/Form/screen_large.jsx
@@ -1,7 +1,7 @@
 import classNames   from 'classnames';
 import PropTypes    from 'prop-types';
 import React        from 'react';
-import UILoader     from 'App/Components/Elements/ui_loader.jsx';
+import UILoader     from 'App/Components/Elements/ui-loader.jsx';
 import Fieldset     from 'App/Components/Form/fieldset.jsx';
 import ContractType from '../../Containers/contract_type.jsx';
 import Purchase     from '../../Containers/purchase.jsx';
