+++ src/javascript/app_2/Modules/Trading/Containers/trade.jsx
@@ -2,7 +2,7 @@ import PropTypes            from 'prop-types';
 import React                from 'react';
 import { CSSTransition }    from 'react-transition-group';
 import { getPropertyValue } from '_common/utility';
-import UILoader             from 'App/Components/Elements/ui_loader.jsx';
+import UILoader             from 'App/Components/Elements/ui-loader.jsx';
 import { connect }          from 'Stores/connect';
 import Test                 from './test.jsx';
 import FormLayout           from '../Components/Form/form_layout.jsx';
