+++ src/sass/app_2.scss
@@ -27,7 +27,7 @@
 @import 'app_2/_common/form/time_picker';
 
 // Components
-@import 'app_2/_common/components/accounts-switcher';
+@import 'app_2/_common/components/account-switcher';
 @import 'app_2/_common/components/allow-equals';
 @import 'app_2/_common/components/calendar';
 @import 'app_2/_common/components/card-list';
