+++ src/sass/app_2/_common/components/account-switcher.scss
